# proto/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "service.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/service/deregister/{registry_id}": {
      "get": {
        "operationId": "Deregister",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbDeregisterResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "registry_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "nodes",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "tags": [
          "Registry"
        ]
      }
    },
...
  },
  "definitions": {
...
    "pbDeregisterResponse": {
      "type": "object"
    },
...
  },
...
}